Stabilize recursive schema
