+++ packages/dds/tree/src/simple-tree/testRecursiveDomain.ts
@@ -10,7 +10,7 @@
  * Currently we do not have tooling in place to test this in our test suite, and exporting these types here is a temporary crutch to aid in diagnosing this issue.
  */
 
-import { SchemaFactoryRecursive } from "./schemaFactoryRecursive.js";
+import { SchemaFactory } from "./schemaFactory.js";
 // This type is referred to by the base type generated by builder.object.
 // Adding the import of it here prevents TypeScript from generating inline includes for it in the d.ts file.
 // If we let TypeScript generate these includes, they use relative paths which break API extractor's rollup.
@@ -18,7 +18,7 @@ import { SchemaFactoryRecursive } from "./schemaFactoryRecursive.js";
 // eslint-disable-next-line @typescript-eslint/no-unused-vars, unused-imports/no-unused-imports
 import { type FieldKind, type NodeKind } from "./schemaTypes.js";
 
-const builder = new SchemaFactoryRecursive("Test Recursive Domain");
+const builder = new SchemaFactory("Test Recursive Domain");
 
 /**
  * To make API-Extractor happy, the base type has to be exported in addition to the actual schema class.
