+++ packages/dds/tree/src/test/simple-tree/schemaTypes.spec.ts
@@ -8,7 +8,7 @@ import assert from "node:assert";
 import { validateAssertionError } from "@fluidframework/test-runtime-utils/internal";
 
 import { TreeValue } from "../../core/index.js";
-import { SchemaFactory, SchemaFactoryRecursive, TreeNode } from "../../simple-tree/index.js";
+import { SchemaFactory, TreeNode } from "../../simple-tree/index.js";
 import {
 	InsertableTreeFieldFromImplicitField,
 	InsertableTypedNode,
@@ -173,7 +173,7 @@ describe("schemaTypes", () => {
 		});
 
 		it("Normalizes recursive schemas", () => {
-			const schemaFactory = new SchemaFactoryRecursive("test");
+			const schemaFactory = new SchemaFactory("test");
 			class Foo extends schemaFactory.objectRecursive("Foo", {
 				x: () => Bar,
 			}) {}
@@ -187,7 +187,7 @@ describe("schemaTypes", () => {
 		});
 
 		it("Normalization fails when a referenced schema has not yet been instantiated", () => {
-			const schemaFactory = new SchemaFactoryRecursive("test");
+			const schemaFactory = new SchemaFactory("test");
 
 			let Bar: any;
 			class Foo extends schemaFactory.objectRecursive("Foo", {
